% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrow_no_na.R
\name{nrow_no_na}
\alias{nrow_no_na}
\title{Number of non missing (NA) values in specified column in dataframe}
\usage{
nrow_no_na(df, colname)
}
\arguments{
\item{df}{Dataframe.}

\item{colname}{String. Column name to filter for non-NA values.}
}
\value{
Integer. Number of rows in df filtered by non-NA values in column colname.
}
\description{
Number of non missing (NA) values in specified column in dataframe
}
\examples{
df <- data.frame(
fruit = c("apple", "banana", "strawberry", "pineapple"),
sugar = c(30, 50, 4, NA)
)
nrow_no_na(df, "sugar")

}
