% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingGITHUB-package.R
\docType{package}
\name{testingGITHUB-package}
\alias{testingGITHUB}
\alias{testingGITHUB-package}
\title{testingGITHUB: Create Beautiful Packages With DARA}
\description{
DARAtest is an R package that provides a comprehensive suite of statistical tests and tools for analyzing data sets, making it easier for users to perform accurate and efficient data analysis.
}
\author{
\strong{Maintainer}: DARAteam \email{dara-team-list@rivm.nl}

}
\keyword{internal}
